lib/igt_kms: drop EDID_LENGTH, replace with EDID_BLOCK_SIZE
EDID_LENGTH is misleading because EDIDs are a variable size (they contain one
or more 128-byte EDID blocks). This commit renames it to EDID_BLOCK_SIZE which
makes it clear users need to call edid_get_size to get the total size.

The declaration has also been moved to igt_edid.

("Size" has been chosen over "length" because it's clearer that it's a number
of bytes, not a number of elements)

Signed-off-by: Simon Ser <simon.ser@intel.com>
Reviewed-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>