tests/kms_3d: Unset all crtcs before the test
On pre-HSW we have separate encoders for the DP and HDMI ports,
but both can't be enabled at the same time. The test fails to
account for that and can thus fail when the kernel rejects
the modeset. We can avoid that by turning off all the ports
beforehand.

v2: rebased (Arek)

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Cc: Simon Ser <simon.ser@intel.com>
Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=105594
Signed-off-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>
Reviewed-by: Simon Ser <simon.ser@intel.com>