tests/kms_plane_multiple: Commit state before remove_fb calls
The kms_plane_multiple@* tests fail on amdgpu with a dmesg-warn due to

"atomic remove_fb failed with -22"
"WARNING: CPU: 6 PID: 183 at drivers/gpu/drm/drm_framebuffer.c:986"

The call to drmModeRmFB essentially creates an atomic commit internally
when we drop the last reference to the FB in IGT. This commit disables
the plane associated with framebuffer.

Since the planes are iterated by increasing index in IGT the last
plane to have its framebuffer removed is the cursor plane on AMDGPU.

On AMDGPU we can't enable the cursor plane without any active hardware
planes (primary, overlay) since there is no dedicated cursor plane.

The commits to release the overlay plane fail since it's the last
remaining plane and the cursor plane is still active.

The fix to the test is to just perform a commit that disables all 3
planes at the same time before calling remove_fb.

This improves performance enables support for drivers that don't
support cursor planes being enabled without a primary or overlay.

Signed-off-by: Nicholas Kazlauskas <nicholas.kazlauskas@amd.com>
Reviewed-by: Leo Li <sunpeng.li@amd.com>