i915/gem_persistent_relocs: Don't call DROP_IDLE in the middle of submitting
To actually DROP_IDLE means that we have to wait for ongoing submission,
and any new concurrently submitted, i.e. it should only be called during
single-threaded submission to ensure the GPU is idle before the new
action.

v2: Also include SHRINK for thrash-all-the-things, and find a dupe in
gem_reloc_vs_gpu.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>