tests/kms_chamelium: check whether EDID blob is non-NULL before trying to read it
test_edid_read sets an EDID on a port, plugs it and reads it back. Sometimes
retrieving the blob fails. Check whether the blob is set at all before trying
to get it. This makes it obvious it's not drmModeGetPropertyBlob failing for
another reason.

Signed-off-by: Simon Ser <simon.ser@intel.com>
Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=106766
Reviewed-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>