prime_busy: Prebind the batch buffer
The test assumes that the dmabuf for the batch buffer has no exclusive
fence. This is only true for the batch as it executes on the GPU; prior
to that we may be in the process of moving it into memory and the
ppGTT which cause the batch to carry an exclusive fence. Keep the test
simple and prebind the batch buffer so that when we poll the dmabuf fd,
we only expect it to be the spinning batch.

Alternatively we could wait until we start writing values into scratch
(i.e. igt_spin_busywait_until_started()). This seemed simpler.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Matthew Auld <matthew.auld@intel.com>
Acked-by: Matthew Auld <matthew.auld@intel.com>