gitlab-ci: rebuild images only on Dockerfile changes
Base images rarely change, there is not much sense in rebuilding it on
every commit. GitLab already has mechanism for detecting such changes.
However, it is only able to prevent rebuilding whenever there is no
changes within the same ref. Since in our CI system, git tag is created
on every series, the mechanism doesn't work.

One possible way to workaround that is to compute a checksum of a
Dockerfile, and rebuilding only if there was no image built with the
same checksum.

V2:
 - Rename gitlab-ci to .gitlab-ci (Arek)
 - Add commit-$CI_COMMIT_SHA

Reviewed-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>
Signed-off-by: Oleg Vasilev <oleg.vasilev@intel.com>