tests/kms_plane_lowres: Skip if no suitable mode is possible
When choosing a low resolution mode to switch to, we search the
connector's mode list for a mode that's small enough that a plane
positioned in the bottom left corner of the screen in the original mode
will be fully offscreen in the new, smaller mode.  However if no
suitable modes are found in the modelist, we blindly fall back to a 10x7
mode without actually checking that the canary plane will now be
offscreen.  If, for example, the original mode was 13x7, then switching
to 10x7 will leave the plane visible onscreen, causing the test to fail.

Before falling back to a 10x7 mode, check whether that's actually small
enough.  If it isn't just skip this subtest.

v2: Restructure the skip condition to make it more obvious why this size
    limitation is important.  (Simon)

Cc: Simon Ser <simon.ser@intel.com>
Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=105581
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>
Reviewed-by: Simon Ser <simon.ser@intel.com>