.gitlab-ci: Don't pull images until they are needed
We were pulling and tagging images locally ahead of time just in case
the next build may need it (i.e. debian-minimal for the fully featured
debian).

Instead of doing that we can specify the registry in Dockerfile's FROM
cluse, so the image is pulled only when we are need it.

Signed-off-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>
Reviewed-by: Petri Latvala <petri.latvala@intel.com>