tests/debugfs_test: Use O_NONBLOCK
Since we are just reading from each debugfs file, any that requires
initial setup may not be fully setup or be left in a state that results
in the file *expectedly* waiting. Use O_NONBLOCK to avoid waits on
external events as we are just peeking for a trivial check that the
files do not explode.

Wrt to the observed failure on ivb, just running the test twice is
enough for it to succeed. It is probably worth investigating exactly
what state is missed on the first pass, but I doubt it has much
relevance outside of this specific test.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Acked-by: Petri Latvala <petri.latvala@intel.com>