meson: Bump required version to 0.47.2
Any distribution that decided to go with 0.47 should also pack the
latest point release.

For us 0.47.2 is important because it fixes the weird issues with using
a 'feature' option as a dependency()'s 'required' parameter, e.g.:

  option('oping',
         type : 'feature',
         description : 'Build test runner with liboping for pinging support')

  liboping = dependency('liboping', required: get_option('oping'))

With liboping not installed and oping option having the default "auto"
value this should execute just fine and liboping.found() should ==
false, but instead we get "ERROR: Native dependency 'liboping' not found".

That behavior is fixed with 0.47.2.

Cc: Petri Latvala <petri.latvala@intel.com>
Signed-off-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>
Reviewed-by: Petri Latvala <petri.latvala@intel.com>