run-tests.sh: Use igt_runner instead of piglit by default
The old behaviour of running everything with piglit with the script is
now behind the flag -p.

When using igt_runner and its pals, -b now takes a blacklist file á la
tests/intel-ci/blacklist.txt.

Additional useful command line flags for igt_runner, like the various
forms of --abort-on-monitored-error or --dmesg-warning-level are not
supported. The main supported form of running tests is igt_runner
directly.

Signed-off-by: Petri Latvala <petri.latvala@intel.com>
Cc: Tomi Sarvela <tomi.p.sarvela@intel.com>
Cc: Arkadiusz Hiler <arkadiusz.hiler@intel.com>
Reviewed-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>