i915/gem_map_gtt: Escape from slow forked GTT access
Beware the slithy t'oves.

Forked GTT access on icl is notoriously slow, so rather than spend an
eternity checking the whole object, check for a completion event after
handling the pagefault. It's is the race of the pagefault vs reset that
we care most about, and we expect the bug to result in the pagefault
being blocked indefinitely, so checking afterwards does not reduce
coverage.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Mika Kuoppala <mika.kuoppala@linux.intel.com>
Reviewed-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>