i915/gem_exec_suspend: Exercise S0 (aka s2idle)
Exercise the first level of suspend, S0. This is basically the same as
our runtime-suspend, we need to put the device to sleep but otherwise
it is left powered up.

Ideally, we would measure the energy consumption in this state.

Maybe /sys/class/power_supply/BAT0/* can be of some value.

References: https://bugs.freedesktop.org/show_bug.cgi?id=111909
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>