meson: Remove redundant install params from configure_file
The 'install' param for configure_file was added in meson 0.50, and
does nothing before that. When using meson >= 0.50 using the install
param gives a warning if the minimum required meson version is lower
than 0.50, because of reasons.

'install : false' is the default anyway when install_dir param is not
set, so remove the redundant and erroneous install param usages.

Signed-off-by: Petri Latvala <petri.latvala@intel.com>
Cc: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>
Cc: Arkadiusz Hiler <arkadiusz.hiler@intel.com>
Reviewed-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>