tests/kms_plane: Don't test every plane on icl+
Only test one plane of each type (HDR, SDR UV, SDR Y) on icl+.
For the purposes of validating pixel formats this should be sufficient
as the programming of each plane of the same type is identical and
as such it's unlikely we'd have different bugs between them.

As before we'll leave the full test set available to be run at
the user's discretion via --extended.

v2: Pimp comment (Petri)

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Reviewed-by: Petri Latvala <petri.latvala@intel.com>