tests/kms_flip: Allow an extra half frame for vblank waits
The vblank tests try to make sure a relative vblank wait of
2 counts completes in ~1-2 frames (depends on how far along the
first frame we are when initiating the wait). The actual
limits we check are 0.9 to 2.1 frames to account for inaccuracies
in the reported dotclock. Unfortunately it seems we either
have even more inaccurate dotclocks or scheduling latencies
and whatnot are pushing that out quite a bit further. So let's
extend the end limit to ~2.6 frames to cut down on the noise a bit.

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=111609
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Reviewed-by: Chris Wilson <chris@chris-wilson.co.uk>