lib/igt_list: Update, clean-up and document igt_list
Our list was something between Wayland and Linux Kernel list
implementations, right in the uncanny valley.

On top of that it falsely claimed that it's a straight copy from the
Wayland project.

Let's make our impl more akin to the kernel one to ease the cognitive
dissonance for the developers working on all those projects.

This patch:
 * mimics the current kernel list interface
 * separates IGT helpers in the source files
 * adds brief explanation and code example for igt-doc
 * introduces igt_list.c as static inlines are not visible in the docs

v2: mimic the kernel instead of wayland (Chris)
    - _head suffix for the sentinel/link struct
    - _entry_ in iterator names that go over the elements
v3: I forgot to merge this in time and there was another call site
    that had to be converted in gem_spin_batch.c

Cc: Petri Latvala <petri.latvala@intel.com>
Cc: Chris Wilson <chris@chris-wilson.co.uk>
Signed-off-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>
Acked-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Petri Latvala <petri.latvala@intel.com>