i915/gem_exec_reloc: Increase spin hostility for faster timeout
Our worst case is blocking on a infinite batch, as the hangcheck allows
it to remain in place indefinitely. This is too slow for CI as it only
fails via a lengthy timeout. However, for our purposes a fail due to
hangcheck is an equally good indicator of the bug, so use a hostile
spinner to provoke hangcheck.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Acked-by: Mika Kuoppala <mika.kuoppala@intel.com>