tests/i915/WA: Mark FF_MODE2 as WO for TGL
On TGL till B0 stepping FF_MODE2(0x6604) register is not readable. But
we need to program this register for Wa_1604555607 for better
performance. Hence to skip the verification attempt on this register
through read, we are marking this register as write only.

As of now I have marked this register as write only for gen12, we can
narrow it down to TGL <= B0, if we can retrieve the stepping of the
platform at IGT.

v2:
  comma is added [Chris]

Signed-off-by: Ramalingam C <ramalingam.c@intel.com>
cc: Tvrtko Ursulin <tvrtko.ursulin@linux.intel.com>
Acked-by: Chris Wilson <chris@chris-wilson.co.uk>