tests/kms_getfb: Add support for GEN12 CCS render compression
GEN12 CCS render compression support is missing from the test.
This causes a SKIP when executing a subtest 'getfb-reject-ccs'.

For TGL the main surface is 4x4 tiles aligned. Therefore, the
pitch for 32bpp is 4*4*32 bytes and the height is 4 rows aligned.

The CCS surface is 64 bytes, which corresponds to 4 tiles on the
main surface. The height of the CCS surface is aligned by 4 rows.

v2: Fix incorrect size and updates on comments (Imre)

Signed-off-by: Mika Kahola <mika.kahola@intel.com>
Reviewed-by: Imre Deak <imre.deak@intel.com>