intel-ci: Drop gem_ctx_switch trivial stress tests
We only use the most trivial of many-context switch stress tests in the
shards. With no verification of behaviour, they only serve to see if the
kernel explodes. We have many, many other context switches with varying
workloads and requirements that _do_ verify expected behaviour and
providing en masse good stress. We also have much more precise
verification of the context switch inside the kernel, making these
simple stress tests redundant.

However, a stress test is a stress test, so keep it around for developer
peace of mind. As usual, if a stress test ever produces a hit, we need
to reduce it down to a new test for CI.

By dropping them we remove over 10 minutes of runtime from the shards,
that afair have never produced a unique warning.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Mika Kuoppala <mika.kuoppala@linux.intel.com>
Cc: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>
Acked-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>