tests/i915_pm_dc: Wait for PSR2 entry
After enable PSR2 we can not check right away if PSR2 is enabled, we
need let some time to PSR2 to be enabled.
For that we have psr_wait_entry(), that is why the previous patch was
reverted as psr_enabled() is not much useful.

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=112411
Cc: Jeevan B <jeevan.b@intel.com>
Cc: Anshuman Gupta <anshuman.gupta@intel.com>
Cc: Arkadiusz Hiler <arkadiusz.hiler@intel.com>
Signed-off-by: José Roberto de Souza <jose.souza@intel.com>
Reviewed-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>