tests/kms_rmfb: Use COMMIT_UNIVERSAL for overlay/cursor planes
Using COMMIT_LEGACY for the cursor plane makes the test fail
because the cursor framebuffer is never handed over to the
kernel, and instead only the BO gets passed in via the
legacy cursor ioctl. Thus the rmfb ioctl will not turn off
the cursor plane since the kernel is using an internal fb
for the cursor. Let's avoid that by treating the cursor as
just another universal plane.

Reviewed-by: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>