intel-ci: Delegate exhaustive fault-injection to the idle runs
i915_module_reload/fault-injection is valuable for its exhaustive trawl
through (hopefully) every path that hits i915_probe_error, unfortunately
such thoroughness comes at a cost. On kbl-guc, it takes almost 3minutes,
and on tgl-y takes over 4 minutes, which is unacceptable for BAT, so leave
it to the shards and idle runs to catch unusual errors on module load.

v2: remove blacklisting to allow module-reload in shards.

Tomi prefers not to have to manually create a runlist for the idle-runs,
so let's take a risk by allowing the module-reload within the shards.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Martin Peres <martin.peres@linux.intel.com>
Cc: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>
Cc: Petri Latvala <petri.latvala@intel.com>
Cc: Tomi Sarvela <tomi.p.sarvela@intel.com>