i915/gem_exec_balancer: Swap bond-chain engines over
Another situation arises where the master is on the same engine as its
signaler. It is submitted to the second ELSP, causing the submission of
its bonded pairs, but it remains blocked. However, if we do not stall
the bonded pair, it will proceed to execute immediately ahead of its
master dependencies.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Acked-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>